requirement.txt
backports.functools-lru-cache==1.6.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
coloredlogs==10.0
cycler==0.10.0
Django==1.11.25
executor==21.3
fasteners==0.15
Flask==1.1.1
Flask-Executor==0.9.3
futures==3.3.0
humanfriendly==4.18
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.4
monotonic==1.5
numpy==1.16.5
opencv-python==4.1.1.26
pandas==0.24.2
pathlib==1.0.1
patsy==0.5.1
plotly==4.2.1
property-manager==2.3.1
pycausalimpact==0.0.15
pyparsing==2.4.2
pyperclip==1.7.0
pyreadline==2.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
retrying==1.3.3
scipy==1.2.2
six==1.12.0
statsmodels==0.10.1
urllib3==1.25.6
verboselogs==1.7
Werkzeug==0.16.0
